Correção de bugs

+++ src/main/java/br/com/compasso/cadastro/service/ClienteServiceImpl.java
@@ -11,35 +11,35 @@ import br.com.compasso.cadastro.repository.ClienteRepository;
 
 @Service
 public class ClienteServiceImpl implements ClienteService {
-	
+
 	@Autowired
 	public ClienteRepository repository;
-	
 
 	@Override
-	public List<Cliente> pesquisaPorNome(String nome) {		
+	public List<Cliente> pesquisaPorNome(String nome) {
 		return repository.pesquisaPorNome("%" + nome + "%");
 	}
 
 	@Override
-	public Optional<Cliente> pesquisaPorId(Long id) {	
+	public Optional<Cliente> pesquisaPorId(Long id) {
 		return repository.findById(id);
 	}
 
 	@Override
-	public Cliente incluir(Cliente cliente) {		
+	public Cliente incluir(Cliente cliente) {
 		return repository.save(cliente);
 	}
-	
+
 	@Override
-	public Cliente alterar(Cliente cliente) {		
+	public Cliente alterar(Cliente cliente) {
 		return repository.save(cliente);
 	}
 
 	@Override
 	public void removerCliente(Long id) {
 		Optional<Cliente> c = repository.findById(id);
-		repository.delete(c.get());
+		if (c.isPresent())
+			repository.delete(c.get());
 	}
 
 }
